Implement GET /profiles/<id> (#223)
